+++ warp/jax.py
@@ -13,13 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# isort: skip_file
-
-from warp._src.jax import get_jax_device as get_jax_device
-
-
 # TODO: Remove after cleaning up the public API.
-
 from warp._src import jax as _jax