+++ warp/render/render_opengl.py
@@ -16,9 +16,13 @@
 # TODO: Remove after cleaning up the public API.
 
 from warp._src.render import render_opengl as _render_opengl
+from warp._src.utils import warn_deprecated_namespace as _warn_deprecated_namespace
 
 
 def __getattr__(name):
     from warp._src.utils import get_deprecated_api  # noqa: PLC0415
 
     return get_deprecated_api(_render_opengl, "wp.render", name)
+
+
+_warn_deprecated_namespace(__name__)